Add a few functions
Add a url-encode function for encoding urls with untrusted query parameters.
Currently, this is a very dumb encoding function.

Add a get-search-url function for building the search url. Currently only
supports using DuckDuckGo but, it's conceivable, that this could be extended to
use any supported search engine.

Add a +duckduckgo-search-url+ constant containing the base search url. We'll add
to this in order to create a complete url for our requests.